refactor(workspaces): split workspace implementations
